% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check.n.mean.R
\name{check.n.mean}
\alias{check.n.mean}
\title{Check if the distribution is close enough to Normal for the Central Limit Theorem to apply.}
\usage{
check.n.mean(y)
}
\arguments{
\item{y}{A variable containing quantitative data.}
}
\description{
Check if the distribution is close enough to Normal for the Central Limit Theorem to apply.
}
