% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check.regression.assumptions.R
\name{check.regression.assumptions}
\alias{check.regression.assumptions}
\title{Diagnostic plots for checking regression assumptions}
\usage{
check.regression.assumptions(model)
}
\arguments{
\item{model}{A linear regression model produced by the \code{lm} function.}
}
\description{
Produces diagnostic plots to aid in checking the LINE regression assumptions.
}
