% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check.regression.conditions.R
\name{check.regression.conditions}
\alias{check.regression.conditions}
\title{Diagnostic plots for checking regression conditions}
\usage{
check.regression.conditions(model)
}
\arguments{
\item{model}{A linear regression model produced by the \code{lm} function.}
}
\description{
Produces diagnostic plots to aid in checking the LINE regression conditions.
}
