% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{civil.rights}
\alias{civil.rights}
\title{Civil Rights Act voting data}
\format{
\subsection{\code{civil.rights}}{

A data frame with 100 rows and 5 columns:
\describe{
\item{state}{The state the senator represents}
\item{vote}{Whether the senator voted "Yea" or "Nay" on the bill}
\item{name}{The name of the senator, including information about party affiliation and tenure}
\item{party}{Party affiliation}
\item{region}{Whether the senator represented a state in the Former Confederacy or not}
}
}
}
\usage{
civil.rights
}
\description{
US senators' votes on the Civil Rights Act.
}
\keyword{datasets}
