% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cramer.v.R
\name{cramer.v}
\alias{cramer.v}
\title{Cramer's V for quantifying the association between two categorical variables}
\usage{
cramer.v(tbl)
}
\arguments{
\item{tbl}{A two-way frequency table comparing two categorical variables.}
}
\description{
Calculates the Cramer's V statistic, which rescales chi-squared to be on a scale from 0 to 1.
}
