% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/z.test.diff.R
\name{z.test.diff}
\alias{z.test.diff}
\title{Two-sample z-test for proportions, with confidence intervals}
\usage{
z.test.diff(
  p1.hat,
  n1,
  p2.hat,
  n2,
  alternative = c("two.sided", "less", "greater"),
  conf.level = 0.95
)
}
\arguments{
\item{p1.hat}{The first sample proportion.}

\item{n1}{The first sample size.}

\item{p2.hat}{The second sample proportion.}

\item{n2}{The second sample size.}

\item{alternative}{The alternative hypothesis, either \code{two.sided} (default), \code{less}, or \code{greater}.}

\item{conf.level}{The confidence level for the confidence interval. Defaults to 0.95 for a 95\% confidence interval.}
}
\description{
Executes hypothesis tests and calculates confidence intervals for the difference of two proportions.
}
