% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/z.test.R
\name{z.test}
\alias{z.test}
\title{One-sample z-test for proportions, with confidence intervals}
\usage{
z.test(
  p.hat,
  n,
  p0 = 0,
  alternative = c("two.sided", "less", "greater"),
  conf.level = 0.95
)
}
\arguments{
\item{p.hat}{The sample proportion.}

\item{n}{The sample size.}

\item{p0}{The value of the population proportion under the null hypothesis. Not needed to calculate confidence intervals..}

\item{alternative}{The alternative hypothesis, either \code{two.sided} (default), \code{less}, or \code{greater}.}

\item{conf.level}{The confidence level for the confidence interval. Defaults to 0.95 for a 95\% confidence interval.}
}
\description{
Executes hypothesis tests and calculates confidence intervals for a proportion.
}
